Created 'GameManager' class

+++ Trivia/Trivia/Helper.h
@@ -14,7 +14,7 @@ public:
 	static Buffer turnStringIntoBuffer(const char* pCh);
 	static char* turnBufferToCharArr(const Buffer& buff);
 
-	static string padNumWith0(int num, int wantedLength = DATA_BYTE_LENGTH);
+	static string padNumWith0(size_t num, int wantedLength = DATA_BYTE_LENGTH);
 	static int charToInt(const unsigned char& digit);
 	static RequestId convertCharsToRequestId(const unsigned char& first, const unsigned char& second);
 	static void setConsoleColor(unsigned int color);
